+++ foundationdb/tests/tokio.rs
@@ -6,16 +6,16 @@ use tokio::runtime::Runtime;
 mod common;
 
 #[test]
-// dropping a future while it's in the pending state should not crash
 fn test_tokio_send() {
-    common::boot();
-
-    let mut rt = Runtime::new().unwrap();
-    rt.block_on(async {
-        do_transact().await;
-        do_trx().await;
+    boot(|| {
+        let mut rt = Runtime::new().unwrap();
+        rt.block_on(async {
+            do_transact().await;
+            do_trx().await;
+        });
     });
 }
+
 async fn do_transact() {
     let db = Arc::new(
         foundationdb::Database::new_compat(None)